* clang_format: (33 commits)
  reformat
  release version 1.8.3
  remove duplicate license file
  fix some merge errors and remaining carriage returns
  add clang format (#26)
  re-add default values for session store from 1.8.3 update
  mozilla is a bad default style
  add a codestyle to remove all those nasty cariage returns
  remove some qt4 compatibility stuff
  ci: Demo3 was deleted
  update to 1.8.3 21.03.2021
  add update script
  ci: don't test Demo2, its build script is broken by design
  replace travis ci with github actions
  Removed unnecessary alias for qt4 (not supported) (#20)
  Added config option to control temp dir (#24)
  Windows ci with msvc and mingw (#23)
  Consistently use qint64 for session time (#17)
  When setting -fsanitize=undefined also set it as linker flag (#18)
  Fix building as dynamic lib with clang (#19)
  ...

# Conflicts:
#	QtWebApp/httpserver/httpconnectionhandler.cpp
#	QtWebApp/httpserver/httpconnectionhandler.h
#	QtWebApp/httpserver/httpcookie.cpp
#	QtWebApp/httpserver/httpcookie.h
#	QtWebApp/httpserver/httplistener.cpp
#	QtWebApp/httpserver/httplistener.h
#	QtWebApp/httpserver/httprequest.cpp
#	QtWebApp/httpserver/httprequest.h
#	QtWebApp/httpserver/httprequesthandler.h
#	QtWebApp/httpserver/httpresponse.cpp
#	QtWebApp/httpserver/httpresponse.h
#	QtWebApp/httpserver/httpserverconfig.cpp
#	QtWebApp/httpserver/httpserverconfig.h
#	QtWebApp/httpserver/httpsession.h
#	QtWebApp/httpserver/httpsessionstore.cpp
#	QtWebApp/httpserver/staticfilecontroller.cpp
#	QtWebApp/httpserver/staticfilecontroller.h
#	QtWebApp/logging/dualfilelogger.h
#	QtWebApp/logging/filelogger.cpp
#	QtWebApp/logging/filelogger.h
#	QtWebApp/logging/logger.cpp
#	QtWebApp/logging/logger.h
#	QtWebApp/logging/logmessage.cpp
#	QtWebApp/logging/logmessage.h
#	QtWebApp/qtwebappglobal.cpp
#	QtWebApp/templateengine/template.h
#	QtWebApp/templateengine/templatecache.cpp
#	QtWebApp/templateengine/templatecache.h
#	QtWebApp/templateengine/templateloader.cpp
#	QtWebApp/templateengine/templateloader.h

+++ QtWebApp/httpserver/httpserverconfig.cpp
@@ -37,6 +37,7 @@ void HttpServerConfig::parseSettings(const QSettings &settings)
 	maxThreads = parseNum(settings.value("maxThreads", maxThreads));
 
 	readTimeout = parseNum(settings.value("readTimeout", readTimeout));
+
 	sslKeyFile = settings.value("sslKeyFile").toString();
 	sslCertFile = settings.value("sslCertFile").toString();
 }
@@ -57,8 +58,9 @@ HttpSessionStoreConfig::HttpSessionStoreConfig(QSettings *settings)
 
 void HttpSessionStoreConfig::parseSettings(const QSettings &settings)
 {
-	expirationTime = parseNum(
-		settings.value("expirationTime", (qulonglong) expirationTime), 1000);
+	expirationTime =
+		parseNum(settings.value("expirationTime", expirationTime), 1000);
+
 	cookieName = settings.value("cookieName", cookieName).toByteArray();
 
 	cookiePath = settings.value("cookiePath", cookiePath).toByteArray();
